+++ src/main/java/com/example/demo/controller/api/app/DANGER_SalesDataAllDeleteApiController.java
@@ -42,10 +42,10 @@ public class DANGER_SalesDataAllDeleteApiController {
 	@ResponseBody // JSONとしてレスポンスするために使う
 	@CrossOrigin
 	public HashMap<String, Object> viewPage(Model model, HttpSession session, @RequestBody PasswordForm password) throws EndPointNotFoundException {
-		if(!password_con.equals(password.getPassword())) throw new EndPointNotFoundException();
+		if(password_con.equals("0") || password_con.equals("") || !password_con.equals(password.getPassword())) throw new EndPointNotFoundException();
 		HashMap<String, Object> data = new HashMap<>();
-		repository.deleteAll();
 		irepository.deleteAll();
+		repository.deleteAll();
 		data.put("result", "ok");
 		return data;
 	}
